Dependencies, Lint, RxJS Imports (#37)
* feat: Update dependencies

Update dependencies on Angular and PrimeNG to avoid deprecation warnings and be compatible wit on Angular and PrimeNG to avoid deprecation warnings and be compatible wit

* refactor: Import RxJS idiomatically to minimize size

Import RxJS idiomatically, follow normal Angular practice. This avoid relying on all of RxJS,
instead relies on just the handful of needed operators.

* fix(selectors): Adjust tslint rules to match 'ao' selector

* fix(package): Correct peer vs devDependencies

Project depends on RxJS as a peerDependency, just as it does Angular itself.

Also, there is no such package as angular/common/http, that is just part of the common package.

* fix: Avoid circular dependencies via barrel

rearrange imports so that there are no circular dependencies via the TypeScript barrel. Also,
remove use of "string" package, instead use regex to strip '/' characters.